refactor: move inline test modules into sibling files
Eight source files were dominated by their inline #[cfg(test)] modules.
Move each test module into a sibling tests.rs so the implementation files
stay readable. Pure code movement, no behavior change.
